+++ src/dotnet-affected/Commands/AffectedRootCommand.cs
@@ -1,49 +1,11 @@
 ﻿using Affected.Cli.Views;
-using DotnetAffected.Abstractions;
-using DotnetAffected.Core;
-using System;
 using System.Collections.Generic;
 using System.CommandLine;
-using System.CommandLine.Binding;
-using System.CommandLine.Invocation;
 using System.CommandLine.Rendering;
 using System.Linq;
 
 namespace Affected.Cli.Commands
 {
-    internal static class InvocationContextExtensions
-    {
-        public static CommandExecutionData GetCommandExecutionData(
-            this InvocationContext context,
-            CommandExecutionDataBinder dataBinder)
-        {
-            if (((IValueSource)dataBinder).TryGetValue(dataBinder, context.BindingContext, out var dataObj) &&
-                dataObj is not null)
-            {
-                return (dataObj as CommandExecutionData)!;
-            }
-
-            throw new InvalidOperationException("Failed to obtain CommandExecutionData from context");
-        }
-
-        public static IAffectedExecutor BuildAffectedExecutor(
-            this CommandExecutionData data)
-        {
-            var options = data.ToAffectedOptions();
-
-            var graph = new ProjectGraphFactory(options).BuildProjectGraph();
-
-            IChangesProvider changesProvider = data.AssumeChanges?.Any() == true
-                ? new AssumptionChangesProvider(graph, data.AssumeChanges)
-                : new GitChangesProvider();
-
-            return new AffectedExecutor(options,
-                graph,
-                changesProvider,
-                new PredictionChangedProjectsProvider(graph, options));
-        }
-    }
-
     internal class AffectedRootCommand : RootCommand
     {
         private static readonly RepositoryPathOptions RepositoryPathOptions = new();
